Refactor line chart renderer
