+++ src/ui-components/chart/cartesian-chart.helper.ts
@@ -477,206 +477,6 @@ export interface RenderTitleOptions {
   titleCssClass?: string;
 }
 
-export const renderLine = (
-  elements: ComponentElements,
-  renderInfo: RenderInfo,
-  component: LineChart,
-  dataset: Dataset,
-  yAxisPosition: string
-): void => {
-  if (!renderInfo || !component) return;
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  let yScale: any = null;
-  if (yAxisPosition === Position.LEFT) {
-    yScale = elements.leftYScale;
-  } else if (yAxisPosition === Position.RIGHT) {
-    yScale = elements.rightYScale;
-  }
-
-  if (component.showLine[dataset.id]) {
-    const lineGen = d3
-      .line<DataPoint>()
-      .defined((p: DataPoint) => p.value !== null)
-      .x((p: DataPoint) => elements.xScale(p.date))
-      .y((p: DataPoint) => yScale(p.value));
-
-    const line = elements.dataArea
-      .append('path')
-      .attr('class', 'tracker-line')
-      .style('stroke-width', component.lineWidth[dataset.id]);
-
-    if (component.fillGap[dataset.id]) {
-      line
-        .datum(Array.from(dataset).filter((p) => p.value !== null))
-        .attr('d', lineGen);
-    } else {
-      line.datum(dataset).attr('d', lineGen);
-    }
-
-    if (component.lineColor[dataset.id]) {
-      line.style('stroke', component.lineColor[dataset.id]);
-    }
-  }
-};
-
-export const renderPoints = (
-  chartElements: ComponentElements,
-  renderInfo: RenderInfo,
-  lineInfo: LineChart,
-  dataset: Dataset,
-  yAxisLocation: string
-): void => {
-  if (!renderInfo || !lineInfo) return;
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  let yScale: any = null;
-  if (yAxisLocation === 'left') {
-    yScale = chartElements.leftYScale;
-  } else if (yAxisLocation === 'right') {
-    yScale = chartElements.rightYScale;
-  }
-
-  if (lineInfo.showPoint[dataset.id]) {
-    const dots = chartElements.dataArea
-      .selectAll('dot')
-      .data(Array.from(dataset).filter((p: DataPoint) => p.value !== null))
-      .enter()
-      .append('circle')
-      .attr('r', lineInfo.pointSize[dataset.id])
-      .attr('cx', (p: DataPoint) => chartElements.xScale(p.date))
-      .attr('cy', (p: DataPoint) => yScale(p.value))
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      .attr('date', (p: DataPoint) => d3.timeFormat('%y-%m-%d')(p.date as any))
-      .attr('value', (p: DataPoint) => {
-        if (p.value !== null) {
-          if (Number.isInteger(p.value)) {
-            return p.value.toFixed(0);
-          }
-          return p.value.toFixed(2);
-        }
-      })
-      .attr('valueType', ValueType[dataset.valueType])
-      .attr('class', 'tracker-dot');
-    if (lineInfo.pointColor[dataset.id]) {
-      dots.style('fill', lineInfo.pointColor[dataset.id]);
-
-      if (
-        lineInfo.pointBorderColor[dataset.id] &&
-        lineInfo.pointBorderWidth[dataset.id] > 0
-      ) {
-        dots.style('stroke', lineInfo.pointBorderColor[dataset.id]);
-        dots.style('stroke-width', lineInfo.pointBorderWidth[dataset.id]);
-      }
-    }
-
-    if (lineInfo.allowInspectData) {
-      renderTooltip(dots, chartElements, renderInfo);
-    }
-  }
-};
-
-export function renderTooltip(
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  targetElements: any,
-  chartElements: ComponentElements,
-  renderInfo: RenderInfo
-): void {
-  const tooltip = chartElements.dataArea.append('svg').style('opacity', 0);
-  const tooltipBg = tooltip.append('rect').attr('x', 0).attr('y', 0);
-  const tooltipLabel = tooltip.append('text');
-  const tooltipLabelDate = tooltipLabel
-    .append('tspan')
-    .attr('class', 'tracker-tooltip-label');
-  const tooltipLabelValue = tooltipLabel
-    .append('tspan')
-    .attr('class', 'tracker-tooltip-label');
-
-  const xSpacing = 3;
-  const ySpacing = 3;
-
-  targetElements
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    .on('mouseenter', (event: any) => {
-      const [x, y] = d3.pointer(event);
-      let tooltipBgWidth = 0;
-      let tooltipBgHeight = 0;
-      // Date
-      const labelDateText = 'date: ' + d3.select(this).attr('date');
-      // labelDateText = x.toString();// debug
-      const labelDateSize = UiUtils.getTextDimensions(
-        labelDateText,
-        'tracker-tooltip-label'
-      );
-      tooltipLabelDate.text(labelDateText);
-      if (labelDateSize.width > tooltipBgWidth) {
-        tooltipBgWidth = labelDateSize.width;
-      }
-      tooltipBgHeight += labelDateSize.height;
-      tooltipLabelDate.attr('x', xSpacing).attr('y', tooltipBgHeight);
-
-      // Value
-      let labelValueText = 'value: ';
-      const valueType = d3.select(this).attr('valueType');
-      const strValue = d3.select(this).attr('value');
-      // strValue += y.toString();//debug
-      if (valueType === 'Time') {
-        const dayStart = window.moment('00:00', 'HH:mm', true);
-        const tickTime = dayStart.add(parseFloat(strValue), 'seconds');
-        const dateValue = tickTime.format('HH:mm');
-        labelValueText += dateValue;
-        tooltipLabelValue.text(labelValueText);
-      } else {
-        labelValueText += strValue;
-        tooltipLabelValue.text(labelValueText);
-      }
-      const labelValueSize = UiUtils.getTextDimensions(
-        labelValueText,
-        'tracker-tooltip-label'
-      );
-      if (labelValueSize.width > tooltipBgWidth) {
-        tooltipBgWidth = labelValueSize.width;
-      }
-      tooltipBgHeight += ySpacing + labelValueSize.height;
-      tooltipLabelValue.attr('x', xSpacing).attr('y', tooltipBgHeight);
-
-      tooltipBgWidth += 2 * xSpacing;
-      tooltipBgHeight += 2 * ySpacing;
-      tooltipLabel
-        .attr('width', tooltipBgWidth)
-        .attr('height', tooltipBgHeight);
-
-      tooltipBg
-        .attr('width', tooltipBgWidth)
-        .attr('height', tooltipBgHeight)
-        .attr('class', 'tracker-tooltip');
-
-      let tooltipPosX = x;
-      let tooltipPosY = y;
-      const tooltipXOffset = 12;
-      const tooltipYOffset = 12;
-      if (x + tooltipXOffset + tooltipBgWidth > renderInfo.dataAreaSize.width) {
-        // move tooltip to left
-        tooltipPosX = x - tooltipBgWidth - tooltipXOffset;
-      } else {
-        // default at the right side
-        tooltipPosX = x + tooltipXOffset;
-      }
-      if (y - tooltipYOffset - tooltipBgHeight < 0) {
-        // down side
-        tooltipPosY = y + tooltipYOffset;
-      } else {
-        // default move to up side
-        tooltipPosY = y - tooltipYOffset - tooltipBgHeight;
-      }
-      tooltip.attr('x', tooltipPosX).attr('y', tooltipPosY);
-      tooltip.transition().duration(200).style('opacity', 1);
-    })
-    .on('mouseleave', () => {
-      tooltip.transition().duration(500).style('opacity', 0);
-    });
-}
-
 export const renderBar = (
   chartElements: ComponentElements,
   renderInfo: RenderInfo,
